✨ components pass for new hooks
